Scripts to Generate a background image with the computer number in.
Set-computerBackground will set the PC Logon background to default image
from a share.

Add-Watermark takes an image and generates the computername text to
embed onto it, creating a new image.

Combined these are used to create a logon screen background for each PC
in a computer lab that has the computer number visible.